pyproject.toml
[project]
name = "propertyprospector"
version = "0.1.0"
description = "Comprehensive tool that scrapes real estate websites, builds a centralized property database, and leverages data analytics to guide your search for the perfect home."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.13.3",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pydoll-python>=1.5.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-snapshot>=0.9.0",
    "ruff>=0.11.0",
    "sqlalchemy>=2.0.39",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.pytest.ini_options]
pythonpath = ["."]
